backend/requirements.txt
# Kairos Trading Bot - Requirements
# Core Dependencies
python-dotenv==1.1.1
requests==2.32.4
colorama==0.4.6

# AI and Language Models
google-generativeai

# FastAPI and Web Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.11.7
python-multipart==0.0.6
aiofiles==23.2.1

# Security and Encryption
cryptography==42.0.8
pydantic[email]==2.11.7

# Database and Vector Storage
supabase==2.4.2
psycopg[binary]==3.2.3
pgvector==0.2.5

# HTTP Client for External APIs
httpx==0.27.0

# AI and Machine Learning
numpy>=1.24.0
scikit-learn>=1.3.0

# PDF Generation for Reports
reportlab==4.0.4